requirements.txt
asgiref==3.7.2
coverage==7.3.2
Django==5.0
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==20.1.0
inflection==0.5.1
iniconfig==2.0.0
mysqlclient==2.2.0
packaging==23.2
pluggy==1.3.0
PyJWT==2.8.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
pytest-factoryboy==2.6.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
